pyproject.toml
[project]
name = "turnero"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cookiecutter>=2.6.0",
    "django>=5.2.4",
    "django-allauth[google,socialaccount]>=65.10.0",
    "django-cotton>=2.1.3",
    "django-tailwind>=4.2.0",
    "honcho>=2.0.0",
]


[tool.djlint]
profile="django"

[dependency-groups]
dev = [
    "django-tailwind[reload]>=4.2.0",
]
